Shun Flag Final
This is the Final Flag for Shun I am going with. I used the ancient Chinese text based on the Flag from Eu4. I redrew the text as coping it is not only lazy but looks bad. No Shun won't have any other flag because the tag will be only for the Monarchy while CHI will be used for republic.