Vietnam
After researching and talking to my Vietnamese friend, without Qing it is almost impossible for Dainam to be used as a name as it is the name the Qing gave them to use.  The names that they could have used are Đại Việt, Nam Việt, or Việt nam. We picked out Vietnam as it's the most well know variation, and most likely with a Viet Nam that is "anti china". I named the leader after my friend.