bruh
The Venetian Republic has been controlled by the Patrician caste for hundreds of years. Their rule is represented by the Patrician Supremacy modifier (below) you get at game start. It prevents you from forming Italy. Removing it requires abolishing the Patrician caste either through the decision (right) or resolving the disaster.

As the world begins to change, springtime of the people-esque conditions will instead kick off a unique disaster where the player will have a choice whether to preserve Patrician rule, reform it, or abolish it. You will be given the choice between a Positivist Intelligentsia agitator employed to reform the caste and a Radical agitator of a non-patrician interest group looking to abolish it entirely. Your support of one over the other will give you access to more characters of agreeing ideologies throughout the disaster while you decide how to deal with the Italian and Greek nationalists,  Aragonese influence in the peninsula, and the tumult that other events bring.