Xijiao stay Xijiao
For Jiangxi/Xijiao, after a while all the characters would become Mahayan, because all the historical people die out.
This is because religion is picked is based off the culture.
The downside to this fix is that every characters in this countries will be Xijiao. This also is placed in 7 different files.
I wish I knew a better way to code this, but this should only effect countries that have Xijiaoas the state religion.