Jiangxi has more resources
Mainly for balance, but there is some history as well. There is a huge lack of iron near this region.
I was able to conquer a region with iron, in my playtest, though some one might have a hard time.
Here is a source though I would like to find one that talks more about the 1800s.
https://www.britannica.com/place/Jiangxi