Fix Batavia Events
Previous checks for Batavia event didn't look for if the events had fired before, and so kept firing ad infinitum. This fix adds a check for a variable to the trigger, and adds a set variable command within the option for the first event, so the Primer can only fire once, and so only trigger the capital event once.